+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 - ComputerManagementDsc:
   - Update psd1 description - Fixes [Issue #269](https://github.com/PowerShell/ComputerManagementDsc/issues/269)
 - Fix minor style issues with missing spaces between `param` statements and '('.
+- SmbServerConfiguration:
+  - New resource for configuring the SMB Server settings.
+  - Added examples for SMB Server Configuration.
+- Minor corrections to CHANGELOG.MD.
 
 ## 7.0.0.0
 
@@ -52,8 +56,6 @@
   - Changed `Test-TargetResource` so that `StartTime` is only compared for
     trigger types `Daily`,`Weekly` or `Once`.
 - Fix minor style issues in statement case.
-- SmbServer:
-  - New resource for configuring the SMB Server settings.
 
 ## 6.5.0.0